Move animation hooks into @react-aria/utils
