+++ packages/@react-aria/utils/src/index.ts
@@ -43,3 +43,4 @@ export {useDeepMemo} from './useDeepMemo';
 export {useFormReset} from './useFormReset';
 export {useLoadMore} from './useLoadMore';
 export {CLEAR_FOCUS_EVENT, FOCUS_EVENT, UPDATE_ACTIVEDESCENDANT} from './constants';
+export {useEnterAnimation, useExitAnimation} from './animation';
